string

+++ main.h
@@ -5,7 +5,7 @@ void print_array(int *array, int o);
 int _printf(const char *format, ...);
 int _print_strin(char *c, int *s, int *i);
 int _putchar(char c);
-int count_size(long int c);
+int count_size(unsigned int c);
 int _bool(const char *format, char *spec, int i);
 int _print_number2(int c);
